+++ include/d/d_event.h
@@ -109,7 +109,7 @@ public:
     fopAc_ac_c* getPt1() { return convPId(mPt1); }
     void setPt2(void* i_actor) { mPt2 = getPId(i_actor); }
     fopAc_ac_c* getPt2() { return convPId(mPt2); }
-    u8 getPreItemNo() { return mItemNo; }
+    int getPreItemNo() { return mItemNo; }
     
     f32 getCullRate() { return mCullFarClipRatio; }
     void setCullRate(f32 ratio) { mCullFarClipRatio = ratio; }
